requirements.txt
# Web Framework
Flask>=3.0.0,<3.1.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.7.0

# ShadowProbe Core Dependencies
pydantic>=2.0.0
python-nmap==0.7.1
dnspython==2.4.2
aiodns==3.1.1
requests>=2.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0
httpx>=0.24.0
selenium>=4.0.0
sqlalchemy>=2.0.0

# Additional Web Dependencies
Werkzeug>=3.0.0,<4.0.0
Jinja2>=3.1.2,<4.0.0
MarkupSafe>=3.0.2

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Optional: Rich for better console output
rich>=13.0.0
tqdm>=4.0.0
whois>=0.9.0

# Added from the code block
# aiodns==3.1.1  # Already included above 

pyOpenSSL>=23.0.0

# Vulnerability Scanning Dependencies
lxml>=4.9.0  # For better XML parsing in some modules
cryptography>=41.0.0  # For cryptographic operations
urllib3>=2.0.0  # For advanced HTTP operations